+++ src/mpl_stereo/__init__.py
@@ -3,4 +3,5 @@ from importlib import metadata
 __version__ = metadata.version(__name__)
 
 from mpl_stereo.AxesStereo import *
+from mpl_stereo.StereoSquare import *
 import mpl_stereo.example_data